Move repetition tests.
